+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 - properly reading API_KEY environment variables for some providers when using Aider modes
 - updated link to telemetry docs
 - user message correctly stored before Aider returns response in case of interrupt or error
+- fixed microphone permission issue on macOS
+- added voice support to browser client
+- added voice settings with microphone selection and additional configuration
+- added Auto Approve lock on project level for new tasks
+- added Auto Approve for Aider modes
 
 ## [0.43.0]